rename block to susi-ligh-login
