+++ test/blocks/susi-light-login/mocks/susi-light-login.html
@@ -1,5 +1,5 @@
 <div>
-  <div class="susi-light-page">
+  <div class="susi-light-login">
     <div>
       <div>linear-gradient(165deg, rgba(251,63,255,1) 0%, rgba(230,255,41,1) 35%, rgba(255,22,22,1) 100%)
       </div>
